refactor: Move/Rename file_handler.go to fs.go
